chore(dapi-grpc): improve lib exports
