+++ Cargo.toml
@@ -15,5 +15,6 @@ members = [
     "packages/feature-flags-contract",
     "packages/dpns-contract",
     "packages/data-contracts",
-    "packages/rs-drive-verify-c-binding"
+    "packages/dapi-grpc",
+    "packages/rs-drive-verify-c-binding",
 ]
