+++ packages/dapi-grpc/src/lib.rs
@@ -1,4 +1,13 @@
-pub mod core;
-pub mod core_grpc;
-pub mod platform;
-pub mod platform_grpc;
+pub mod core {
+    #[cfg(feature = "core_v0")]
+    pub mod v0 {
+        include!("core/proto/org.dash.platform.dapi.v0.rs");
+    }
+}
+
+pub mod platform {
+    #[cfg(feature = "platform_v0")]
+    pub mod v0 {
+        include!("platform/proto/org.dash.platform.dapi.v0.rs");
+    }
+}
