--- packages/dapi-grpc/src/proto2rs.rs
@@ -1,8 +1,0 @@
-mod generate;
-
-fn main() {
-    print!("Generating protobuf definitions... ");
-
-    crate::generate::generate().expect("protobuf generation failed");
-    println!("DONE");
-}
